loadconfig/loadtracing and run signal interupted func moved to application package
