+++ internal/infra/application/run_signal_interruption_func.go
@@ -1,4 +1,4 @@
-package infra
+package application
 
 import (
 	"context"
